bd/requirements.txt
asgiref==3.5.2
attrs==21.4.0
awsebcli==3.20.3
bcrypt==3.2.2
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==37.0.2
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.3
importlib-metadata==4.11.4
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.11.0
pathspec==0.9.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycparser==2.21
PyGreSQL==5.2.4
PyJWT==2.4.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.8.0
